+++ src/main/java/net/rubyeye/xmemcached/command/TextCommandFactory.java
@@ -200,7 +200,7 @@ public final class TextCommandFactory implements CommandFactory {
 	 * .String, byte[], int, net.rubyeye.xmemcached.command.CommandType)
 	 */
 	public final Command createIncrDecrCommand(final String key,
-			final byte[] keyBytes, final int amount, int initial, int exptime,
+			final byte[] keyBytes, final long amount, long initial, int exptime,
 			CommandType cmdType, boolean noreply) {
 		return new TextIncrDecrCommand(key, keyBytes, cmdType,
 				new CountDownLatch(1), amount, initial, noreply);
